+++ src/cspfj/heuristic/Lexico.java
@@ -9,7 +9,7 @@ public final class Lexico extends AbstractStaticValueHeuristic {
         super(problem);
     }
 
-    public float getScore(final Variable var, final int index) {
+    public double getScore(final Variable var, final int index) {
         return index;
     }
