+++ src/cspfj/heuristic/ValueHeuristic.java
@@ -4,6 +4,6 @@ import cspfj.problem.Variable;
 
 public interface ValueHeuristic {
 	int selectIndex(Variable variable) ;
-    float getScore(final Variable variable, final int index) ;
+    double getScore(final Variable variable, final int index) ;
     void compute() ;
 }
